Move tests into test_electrum.py
